refactor: 코드 리뷰 반영
- @Transactional을 jakarta에서 org.springframework으로 변경 (TransactionSynchronizationManager 호환)
- FileValidator.validateMimeType()이 감지된 MIME 타입을 반환하도록 변경, FileUploadService에서 클라이언트 선언값 대신 Tika 감지값을 메타데이터에 저장
- FileValidator 상수 및 검증 메서드를 package-private으로 변경, FileValidatorTest를 동일 패키지(com.gathering.file.application)로 이동
- deleteProfileImage에서 profileImageUrl null 체크 대신 메타데이터 존재 여부로 조기 반환
- LocalStorageService.extractExtension()에서 빈 확장자 시 예외 throw로 방어 코드 강화
